+++ CHANGELOG.md
@@ -1,6 +1,6 @@
-##### Wed Aug 03 2016 Sundeep Anand <suanand@redhat.com>
-* add transplatform, locales settings pages (read-only)
-* add alembic and settings page
+##### Thu Aug 04 2016 Sundeep Anand <suanand@redhat.com>
+* add languages, transplatform, relstreams settings pages (read-only)
+* add alembic and settings page/models
 * rename project to transtats, fix ui and migrate to PostgreSQL
 * add features page, zanata api, view, manager and model
 * link travis and readmedocs